// desktop/src/open-api/model/systemTaskPagedRequestFilter.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SystemTaskPagedRequestFilter { 
    /**
     * Contains two keys: operation of type FilterOperation and value which can a different type depending on the field.
     */
    type?: object;
    /**
     * Contains two keys: operation of type FilterOperation and value which can a different type depending on the field.
     */
    ranBy?: object;
    /**
     * Contains two keys: operation of type FilterOperation and value which can a different type depending on the field.
     */
    startedAt?: object;
    /**
     * Contains two keys: operation of type FilterOperation and value which can a different type depending on the field.
     */
    endedAt?: object;
}